requirements.txt
fastapi
uvicorn
pydantic
openai
playwright
httpx
pdfplumber
pandas
numpy
matplotlib
python-dotenv
reportlab